pyproject.toml
[tool.poetry]
name = "test_fastapi"
version = "0.1.0"
description = "This is a simple note-taking service"
authors = ["Babayev Ruslan <rusl.babae@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.27.1"
sqlalchemy = {version = "^2.0.27", extras = ["asyncio"]}
loguru = "^0.7.2"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
fastapi = "0.100.0"
asyncio = "^3.4.3"
pydantic = "^2.6.4"
python-multipart = "^0.0.9"
pyjwt = "^2.8.0"
python-keycloak = "^3.9.1"
fastapi-cache2 = "^0.2.1"


[tool.poetry.scripts]
launch_notes_api = "test_fastapi.__main__:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
